[#55] fix: src/@types 적용
