+++ src/components/Loading/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Loading() {
+  return <div>로오딩</div>;
+}
+
+export default Loading;
