+++ src/components/Profile/CardList/Card.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import Image from "next/image";
 import styled from "styled-components";
-import { IGetCardDataContent } from "src/pages/profile";
 
 interface ICard {
   card: IGetCardDataContent;
